WIP: Add CredentialsManager
Still need to finish this off, but waiting until maybe this is supported
in aws-amplify master branch. Could either a) take this approach and listen
on Hub for credential changes or b) add federated identity support to AuthClass.

Note that the event broadcasted via Hub does not currently have all the info
necessary to set AWS.config.credentials - that is filtered out via essentialCredentials.